requirements.txt
langchain
boto3
pdfplumber
pandas
openpyxl
python-dotenv
langchain_aws
langgraph
python-dotenv
datetime
pydantic
botocore
scikit-learn
nltk
numpy
reportlab
matplotlib
seaborn
streamlit
plotly
